requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
Faker==30.8.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
